+++ docs/source/release_notes/release-notes.rst
@@ -48,6 +48,10 @@ develop
     *    - Type
          - Change
 
+    *    - |improved|
+         - Range scan iterators returned by `ProfilingKeyValueService` now log `kvs-slow-log` warnings when the iterator is accessed if applicable.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/2663>`__)
+
     *    - |fixed| |metrics|
          - ``MetricsManager`` now logs failures to register metrics at ``WARN`` instead of ``ERROR``, as failure to do so is not necessarily a systemic failure.
            Also, we now log the name of the metric as a Safe argument (previously it was logged as Unsafe).